Fixes: When searching from a fullpost, you now get a new set of results on the new search term
Doing a search from the "pick-a-date" component did not always get the filters attached to the query.

Now picking a new month or year in the "pick-a-date" component sets the day to the first, so you don't _have_ to pick a date before doing a search.

Fixed an error where you could get a weird breaking of the app because a response was not checked if existed before looking into the object.

Fixed a bug where the values of the startDate and endDate were not reset when returning to the frontpage, thus messing up the timeline component's "go look at this"-link