Facets are now called separately. If facets are opened when we do a new search, they're called, otherwise first when the "show filters" are opened. This gives a little loadtime, but it's fine.
Back button now goes to the last search, wherever you might have gone, or back to the frontpage if there was no search.

Videoplayer new respons to the 1002 error from Kaltura, throwing an error and telling the user they cant access this recording.

Fixed a bug with filters not being applied right in the timeSearchStore.

Fixed a few background accessability issues.

Removed a few console.logs