Facets should now open nicely (no jumping in the animation)
Filters buttons should now stay in a pretty decent place on all screen-sizes.

The cut into the header on FullRecord should now be consistent on all devices.

Added a small fix to the player sizes.

Added a small fix so when you apply time filters, it actually scrolls down to the results instead of a weird jump to nothing.